Add Q_GADGET Wrapper for QQmlProperty
This enable exposing a QQmlProperty as a value type to QML code.
Only object and name are exposed as a properties to keep it simple.

We might want later to expose the property of a Binding or a Behavior
as a property for complex usecases. This permits exposing it as one
self contained value instead of a pair of unrelated object and name.

Task-number: QTBUG-70964
Task-number: QTBUG-73892
Change-Id: I3a46212446f43f3a7e301943cb49d3a48c377de3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>