Fix printing of garbled UTF-8 characters from QML verify() statements
Messages that contain QtPositioning.coordinate are including the degree
unicode symbol U+00B0. These were printed as garbage from QML tests that use
the verify() function. This was caused by the message argument being encoded
using toLatin1() on its conversion to const char *.

This caused both garbled text and invalid XML output in the test runs, which
in turn was causing coin's agent to fail while deserialising the test XML log.

This was fixed by changing from toLatin1() to toUtf8(). Since testlib is
receiving plain const char * arguments for the "description" argument, it was
deemed correct to standardize on UTF-8 as the proper encoding for it, to avoid
losing any information (QTBUG-81474). When it's output to the console it can
be re-encoded depending on the system's locale, of course.

Task-number: QTBUG-81474
Fixes: QTBUG-80527
Change-Id: I0f1fdb7d414d9dfd30d4638b00088c9502c339bd
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>