Native style: copy QStyle from widgets, and implement ground work for creating native styles
This is the initial patch for adding native style support to controls2.

The gist of the patch is that it copies QStyle from widgets (including the
mac plugin style), remove all notions of widgets, tweak it as needed, and
make it compile inside the qtquickcontrols2 repository as a separate
QML plugin (QtQuick.NativeStyle).

The "new" QStyle is then used to draw primitives onto nine-patch-images that can
be rendered by the scene graph. Each such primitive/image will be wrapped by a
QQuickStyleItem (which is a subclass of QQuickItem). E.g a button background will
be implemented by QQuickStyleItemButton.cpp. This item can then be placed anywhere
in the QML code to draw a native-looking button background.

Controls2 has its own styling API, where a style consist of a set of QML files
that extend template controls written in C++. To enable native styling, we simply
follow the exact same approach; We create a style folder per platform alongside
the other styles in controls2. Each style will contain a set of qml files for the
controls (Button.qml, Slider.qml, etc), and inside each control, the style items
will be used to draw the default delegates. Since it's likely that each such
qml file will be implemented equal in all the desktop styles (but not always), the
native style plugin contains a set of default controls (e.g DefaultButton.qml) that
a platform style can choose to inherit from and extend as needed.

Included is also an example called "DesktopGallery" that can be used to
test and view the controls.

Change-Id: I8b45aa7d493930f552d3ad2e3e3e7184129a9d6c
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>