QQmlInstanceModel: refactor recycling signals to base class
Now that QQmlDelegateModel has an API that handles reusing
delegate items (*), we should also move the related signals
inside it to be consistent. This will also remove the need
to cast the model type in the views before connecting.

This patch will also remove warnings that stems from
QQuickListView trying to connect to the reuse signals
when the model is not a QQmlDelegateModel.

*: E.g: virtual ReleaseFlags release(QObject *object,
    ReusableFlag reusableFlag = NotReusable) = 0;

Fixes: QTBUG-81257
Change-Id: Ia8a8f0d68e6ef7edc6c45b414121aaa77632bcd3
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>