Update foreign types list generation to use ninja dep files
This patch updates the generation of the foreign type list to use ninja
dep files in order to only run if on of the calculated dependencies
change.

This patch also extends generation of the cpp file from qmltypesregistrar
to use dep files so it runs if one of the dependent foreign metatypes.json
is modified.

Change-Id: I31c0dcf7cffea0c5277347f62115d9059244939a
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>