Copy Qt Quick Designer-related files also for static builds
Make sure that the 'designer' directory is copied to the build
directory, or installed, also for static builds.

This reverts the !static condition introduced to fix a build
breakage in 01d076d0e81e5 . Anyhow, I couldn't reproduce
the original problem (QTBUG-51708), nor is it clear why it
was caused by the static build.

Fixes: QTBUG-75682
Change-Id: I47af3907e464d4fd9e9965ac28545b7350f450ee
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Shaw <andy.shaw@qt.io>