Add additional properties to FontLoader to help disambiguate
The FontLoader only returns the family name, and there is no
way to disambiguate fonts within the same family. This
change adds the option of getting other properties from the
loaded font to help with this.

[ChangeLog][QtQuick] Added FontLoader.font property which
is filled with a font query that matches the loaded font.
This can be used to disambiguate when multiple fonts in the
same family  are loaded in the same application.

Fixes: QTBUG-68829
Change-Id: If05e85a1ca8d93b93100345cb2d05cd1ad6c7549
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>