Prospective workaround for accidental c++11 build with 10.14
When switching to macOS 10.14 and the xcode we install there,
mysteriously qtbase switches from c++17 to c++11 when compiling the
tests, which breaks this particular test as it includes
qv4arraydata_p.h, which in turn requires C++14's std::aligned_storage_t.

Once we enforce C++17 in qtbase this should not be needed anymore, but
we can only enforce that once we've switched to macOS 10.14.

Change-Id: Iaa2b55dfd2bfd23f908a211ca90255ef522bc6ff
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>