Stabilize tst_qqmltimer::restart()
Double the intervals. Some operating systems are sloppier with
scheduling and we apparently trigger the timer a third time during the
wait interval. We waited 600ms + 300ms = 900ms and expected the 500ms
timer to hit only once before the restart. That was probably too tight.

We also lower the second wait time by 100ms to allow for some more time
skew, and check all intermediate states of the trigger counter.

Change-Id: I2f83ffb0e05f1939d3525cd649322a4331986c3c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Maximilian Goldstein <max.goldstein@qt.io>