Account for when a touch event is synthesized by Qt as a mouse event
When a control is on a Flickable with a pressDelay then any press events
sent from a touch device will be replayed as mouse events due to the
delay. As a result we cannot depend on the fact that we got the first
press as a touch event when checking if the id matches before accepting
it. So we need to keep the previous pos when it is a synthesized mouse
event so we can ensure the release is also accepted.

Fixes: QTBUG-77202
Change-Id: I6f5d8506bd803daf834093e8fd412504150c4ca6
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>