Fix QQmlObjectModel::destroyingItem emission
The signal is used to tell the view that the item is definitely going
away. For the "view" based QQmlObjectModel that is almost never really
the case, except - oddly - for clear(). The view is typically a
QQuickItemView, which casts the item to a QQuickItem and calls
setParentItem(nullptr). That in turn is caught by QQuickContainer, which
calls remove() on the QQmlObjectModel. That is why remove() can't emit
destroyingItem().

Amends 6d0a453f41d304239285d64b06612c36922be701

Change-Id: I5d82def872550744b947b4b53447647327e03f67
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>