QQuickItemView: do not set parent to nullptr upon destruction
Setting a delegate item's parent to nullptr will make any
bindings to the parent inside the item invalid, and
warnings will as such be printed (first seen after
8c72e634b3b0eacbfdee883bfc34994d3c19ed77).

I assume the reason for setting the (visual) parent to nullptr
is to immediately hide it. So to avoid trigger any bindings, we
instead just cull if from the scene graph.

It's also dubious why a delegate should have bindings to
its parent in the first place, since what ends up being the
parent is an implementation detail, and probably not be the item
the user expects it to be.

Fixes: QTBUG-81976
Change-Id: I7bd8ab91461504b6e79d4aa2ab832be087245c3e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>