Add a null check to QQuickWindowPrivate::renderer
Afaik there currently is no case in a regular execution where the
renderer would actually be null, but for the sake of correctness add a
null check as this seems to be the only code location that actually
assumes the renderer to be non-null.

Additionally, in order to en-/disable custom render modes at runtime,
GammaRay actually needs to recreate the renderer and therefore resets
the renderer to null in order for QQuickWindowPrivate::syncSceneGraph to
recreate it. Thus we need a null check to make sure we don't crash in
the short time frame where the renderer is null.

Change-Id: Ief5c405f3bc8725d55e22cd33f2164830764e33d
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>