Windows: Do look for <plugin>d.dll as a fallback
Do look for d.dll as name of the library, even if Qt was built in
release mode. This mitigates the change we did in Qt 5.14.0, where
we switched the way our MinGW builds are configured from
-debug-and-release to -debug -force-debug-info. But it is also
more consisten in how we handle the release build, and macOS
dylib loading.

Fixes: QTBUG-81021
Change-Id: I73cd11941d86a732b92a3479d47b6e83a839994f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>