ToolTip: fix timeout not activating if opened during exit transition
When a tooltip with an exit transition and a timeout is made visible
during its exit transition, it would end up not timing out. This is
because the change in the tooltip's visibility is what usually
kicks off the timeout timer, but if there is no change in visibility
(because the exit transition hadn't finished, so the tooltip was
still visible), this won't happen.

Fix the issue by calling startTimeout() in
QQuickToolTip::setVisible().

Change-Id: I5f95a698e48f376c7597558572ac91625a276e53
Fixes: QTBUG-81935
Reviewed-by: Andy Shaw <andy.shaw@qt.io>