HeaderView: set implicitSize on the style items
By setting an implicit size, the user don't need to set a
width or height on a HeaderView himself, but it will
get the default size recommended by the style.

By doing it the way it's done in the patch we achieve the following:

1. A HeaderView will by default be resized to be the same size
    as the delegate.
2. If the application sets a size on HeaderView it that is larger
    than the implicit size of the delegate, the delegate
    will be resized to have the same size (effectively
    filling out the free space in the header).
3. If the size of HeaderView is smaller than the implicit
    size of the delegate, the delegate will simply be clipped.
    (effectivly saying that the implicitSize of the delegate is
    also it's minimum size). If this is not acceptable for the
    application, it will need to use a custom delegate.

Since a HeaderView delegate is a component and not an item, it
should not be a part of the sanity checks we do to avoid using
internal IDs. Hence we blacklist until we have a better way of
handling such cases.

Task-number: QTPM-1300
Change-Id: I30ca3e13ce5e1371b60f5c4ecf742a7d7e794a36
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>