TableModel: Allow a double to be added to a field seen as int previously
Since the type for a TableModelColumn can be seen as an int when it is
in the first row, it is still possible that it is representing a double
value in other rows. Therefore it should be allowed to add/change a row
that would display a double for that column.

Change-Id: I994a8ead595f836a20a4e82fbf94953b1aa4b7da
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>