Don't replace all characters below 0x20 with spaces as they may be valid
Some fonts will still render something for characters below 0x20 so
these should not be replaced with a space and should just be left as is.
It will not be rendered as a box if the dontPrint flag is set so we can
trust the renderer to do the right thing here.

Change-Id: Ie091c14713d8c2948c82b1991c295d80bc35dfde
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>