Remove comments about renaming and removing properties for Qt6
For the record, my argument for not acting upon these comments was that
they seems to be missing a key issue:
OrientationSensor does not offer a superset of the features of
QQuickScreen::orientation. Therefore, it cannot just replace
QQuickScreen::orientation: An OrientationReading is the output from the
sensor itself regardless of any settings that was made in the OS to for
instance lock the UI to landscape only. And this seems to be the point
of the QScreen::orientation: It's supposed to be the "virtual/locked"
orientation we get when we ask the OS (which might not reflect the true
physical orientation, since it might be overridden by the OS due to e.g.
landscape locking).

The original author agreed with this argument, and agreed that the
comments can just be removed.

Change-Id: I2df0ae9debc6b6177f0e4ee6efb60c03da726e15
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>