Make Connections object qmlWarning follow a logging category
Connections{} emits a warning if an old syntax is used.

Unfortunately that syntax is not available in Qt5.12. Downstream needs
to support both at once, which means we have code constantly hitting
this warning. We can't even #ifdef the QML code.

By introducing a logging category we can still have the warning but
provides some option for downstream to temporarily disable through an
existing mechanism until older Qt is not supported.

Task-number: QDS-2007
Change-Id: Idd34c18ae3e8b3e53c430fab8229bb9812370d0e
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>