Avoid unloading plugins which register types to prevent crashes
Plugins could register an arbitrary number of types in arbitrary places
with qRegisterMetaType. This would lead to crashes when the plugin is
unloaded, as the type registry would hold stale references to their
QMetaTypes. By not unloading the plugin, the metatypes are kept safe in
memory.

[ChangeLog][QML][Important Behavior Changes] When writing a
QQmlExtensionPlugin which procedurually registers types, you should
overload the newly introduced unregisterTypes function, if you want to
support unregistration with qmlClearEnginePlugins. Additionally, the
destructor of your plugin should not contain any logic.

Change-Id: I63b7a153f82be7996dbfca4d36f279c351dc6b9e
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>